finish a2 combining extractor implementation

+++ test/test_data_filter.cc
@@ -389,7 +389,7 @@ void TestDataFilter::test_multiwordfilter()
         QVERIFY(process_data(&mf, 0xcd));
         QCOMPARE(extract(&mf, MultiWordFilter::CacheA), (u64)(0xda));
         QCOMPARE(extract(&mf, MultiWordFilter::CacheD), (u64)(0xcb));
-        QCOMPARE(get_extract_bits(&mf, MultiWordFilter::CacheA), (u8)8);
+        QCOMPARE(get_extract_bits(&mf, MultiWordFilter::CacheA), (u16)8);
     }
 
     // slow MultiWordDataFilter
